chore: Move testutil from internal to root
