+++ coderd/provisionerjobs_internal_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/coder/coder/coderd/database"
 	"github.com/coder/coder/coderd/database/databasefake"
 	"github.com/coder/coder/codersdk"
-	"github.com/coder/coder/internal/testutil"
+	"github.com/coder/coder/testutil"
 )
 
 func TestProvisionerJobLogs_Unit(t *testing.T) {
