+++ testutil/duration_windows.go
